Added option 'Nullify_MissingK' to parameters
Added an option 'Nullify_MissingK' to specify if extra frequencies, K,
present in the simulation but not in the input Forcing Energy Cascade,
i.e., "missing K", need to be nullified or allowed its initial state to
decay normally.
This allows, for instance, to only inject energy to the large scales and
let the system decay.